forensics/julens_undergang/src/requirements.txt
pwntools
pyasyncore
pycryptodome
pyOpenSSL
tlslite-ng
